feat: add resume tab & fix resume input
